Escape autosave functionality into its own composable
